Completed 4.21 and 4.22

+++ controllers/blogs.js
@@ -1,7 +1,5 @@
-const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
 
 
 blogsRouter.get('/', async (request, response) => {
@@ -22,19 +20,10 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 
-const getTokenFrom = request => {
-
-}
-
-
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token invalid' })
-    }
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     const blog = new Blog({
         title: body.title,
@@ -58,7 +47,13 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+    }
+    // await Blog.findByIdAndDelete(request.params.id)
     // console.log(deletedBlog)
     // if (deletedBlog) {
     //     response.status(204).end()
